Fix add user & rooms

+++ Demo-SpringBoot-SOA-BE/src/main/java/com/example/demospring/controller/RoomController.java
@@ -4,12 +4,14 @@ import com.example.demospring.model.Room;
 import com.example.demospring.service.RoomService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
 import java.util.List;
 
 @CrossOrigin(origins = "http://localhost:4200")
 @RestController
 @RequestMapping("/api/rooms")
 public class RoomController {
+
     private final RoomService service;
 
     public RoomController(RoomService service) {
@@ -17,26 +19,30 @@ public class RoomController {
     }
 
     @GetMapping
-    public List<Room> getAll() { return service.getAll(); }
+    public List<Room> getAll() {
+        return service.getAll();
+    }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Room> getById(@PathVariable Integer id) {
+    public ResponseEntity<Room> getById(@PathVariable Long id) {
         return service.getById(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
     }
 
     @PostMapping
-    public Room create(@RequestBody Room room) { return service.save(room); }
+    public Room create(@RequestBody Room room) {
+        return service.save(room);
+    }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Room> update(@PathVariable Integer id, @RequestBody Room room) {
+    public ResponseEntity<Room> update(@PathVariable Long id, @RequestBody Room room) {
         room.setId(id);
         return ResponseEntity.ok(service.save(room));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Integer id) {
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
         service.delete(id);
         return ResponseEntity.noContent().build();
     }
