+++ Demo-SpringBoot-SOA-BE/src/main/java/com/example/demospring/repository/RoomRepository.java
@@ -3,5 +3,5 @@ package com.example.demospring.repository;
 import com.example.demospring.model.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomRepository extends JpaRepository<Room, Integer> {
+public interface RoomRepository extends JpaRepository<Room, Long> {
 }
